requirements.txt
Django==3.2.9
psycopg2-binary==2.9.2
djangorestframework==3.12.
channels==3.0.4
channels-redis==3.3.14
